Form validation by javascript
CheckValidity function was working at the start but no longer does. For some reason the inputs [firstname], [lastname], [telephone] are no longer starting the function. I added checkValidity to the submitForm function which gives the form the proper functionality that it needs. There is no downside to this work around since anything that is typed in these 3 fields are valid.